requirements.txt
notion-client==1.0.0
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.2
